requirements.txt
Flask==0.10.1
Flask-Redis==0.1.0
Flask-LDAPConn==0.6.7
Pillow==3.0.0
